requirements.txt
python-telegram-bot>=21.0
google-genai
python-docx
pandas
openpyxl
yfinance
Flask